+++ pkg/auth/rights/auth_info.go
@@ -18,6 +18,7 @@ import (
 	"context"
 
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
+	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 )
 
@@ -46,4 +47,22 @@ func AuthInfo(ctx context.Context) (authInfo *ttnpb.AuthInfoResponse, err error)
 		return nil, err
 	}
 	return authInfo, nil
+}
+
+var errUnauthenticated = errors.DefineUnauthenticated("unauthenticated", "unauthenticated")
+
+// RequireAuthentication confirms if the authentication information within a context contains any rights, if so,
+// the request is considered to be authenticated.
+func RequireAuthentication(ctx context.Context) error {
+	log.FromContext(ctx).Debug("Authenticate request")
+	authInfo, err := AuthInfo(ctx)
+	if err != nil {
+		log.FromContext(ctx).WithError(err).Debug("Failed to validate authentication information")
+		return errUnauthenticated.WithCause(err)
+	}
+
+	if authInfo.GetAccessMethod() == nil && len(authInfo.GetUniversalRights().GetRights()) == 0 {
+		return errUnauthenticated.New()
+	}
+	return nil
 }