+++ pkg/devicerepository/grpc.go
@@ -60,13 +60,12 @@ func (dr *DeviceRepository) ensureBaseAssetURLs(models []*ttnpb.EndDeviceModel) 
 
 const defaultLimit = 1000
 
-func requireApplication(ctx context.Context, ids *ttnpb.ApplicationIdentifiers) error {
-	return rights.RequireApplication(ctx, ids, ttnpb.Right_RIGHT_APPLICATION_DEVICES_READ)
-}
-
 // ListBrands implements the ttnpb.DeviceRepositoryServer interface.
-func (dr *DeviceRepository) ListBrands(ctx context.Context, req *ttnpb.ListEndDeviceBrandsRequest) (*ttnpb.ListEndDeviceBrandsResponse, error) {
-	if err := requireApplication(ctx, req.GetApplicationIds()); err != nil {
+func (dr *DeviceRepository) ListBrands(
+	ctx context.Context,
+	req *ttnpb.ListEndDeviceBrandsRequest,
+) (*ttnpb.ListEndDeviceBrandsResponse, error) {
+	if err := rights.RequireAuthentication(ctx); err != nil {
 		return nil, err
 	}
 	if req.Limit > defaultLimit || req.Limit == 0 {
@@ -94,8 +93,11 @@ func (dr *DeviceRepository) ListBrands(ctx context.Context, req *ttnpb.ListEndDe
 var errBrandNotFound = errors.DefineNotFound("brand_not_found", "brand `{brand_id}` not found")
 
 // GetBrand implements the ttnpb.DeviceRepositoryServer interface.
-func (dr *DeviceRepository) GetBrand(ctx context.Context, req *ttnpb.GetEndDeviceBrandRequest) (*ttnpb.EndDeviceBrand, error) {
-	if err := requireApplication(ctx, req.GetApplicationIds()); err != nil {
+func (dr *DeviceRepository) GetBrand(
+	ctx context.Context,
+	req *ttnpb.GetEndDeviceBrandRequest,
+) (*ttnpb.EndDeviceBrand, error) {
+	if err := rights.RequireAuthentication(ctx); err != nil {
 		return nil, err
 	}
 	response, err := dr.store.GetBrands(store.GetBrandsRequest{
@@ -115,8 +117,11 @@ func (dr *DeviceRepository) GetBrand(ctx context.Context, req *ttnpb.GetEndDevic
 }
 
 // ListModels implements the ttnpb.DeviceRepositoryServer interface.
-func (dr *DeviceRepository) ListModels(ctx context.Context, req *ttnpb.ListEndDeviceModelsRequest) (*ttnpb.ListEndDeviceModelsResponse, error) {
-	if err := requireApplication(ctx, req.GetApplicationIds()); err != nil {
+func (dr *DeviceRepository) ListModels(
+	ctx context.Context,
+	req *ttnpb.ListEndDeviceModelsRequest,
+) (*ttnpb.ListEndDeviceModelsResponse, error) {
+	if err := rights.RequireAuthentication(ctx); err != nil {
 		return nil, err
 	}
 	if req.Limit > defaultLimit || req.Limit == 0 {
@@ -143,8 +148,11 @@ func (dr *DeviceRepository) ListModels(ctx context.Context, req *ttnpb.ListEndDe
 var errModelNotFound = errors.DefineNotFound("model_not_found", "model `{brand_id}/{model_id}` not found")
 
 // GetModel implements the ttnpb.DeviceRepositoryServer interface.
-func (dr *DeviceRepository) GetModel(ctx context.Context, req *ttnpb.GetEndDeviceModelRequest) (*ttnpb.EndDeviceModel, error) {
-	if err := requireApplication(ctx, req.GetApplicationIds()); err != nil {
+func (dr *DeviceRepository) GetModel(
+	ctx context.Context,
+	req *ttnpb.GetEndDeviceModelRequest,
+) (*ttnpb.EndDeviceModel, error) {
+	if err := rights.RequireAuthentication(ctx); err != nil {
 		return nil, err
 	}
 	response, err := dr.store.GetModels(store.GetModelsRequest{
@@ -165,16 +173,23 @@ func (dr *DeviceRepository) GetModel(ctx context.Context, req *ttnpb.GetEndDevic
 }
 
 // GetTemplate implements the ttnpb.DeviceRepositoryServer interface.
-func (dr *DeviceRepository) GetTemplate(ctx context.Context, req *ttnpb.GetTemplateRequest) (*ttnpb.EndDeviceTemplate, error) {
-	if err := requireApplication(ctx, req.GetApplicationIds()); err != nil {
+func (dr *DeviceRepository) GetTemplate(
+	ctx context.Context,
+	req *ttnpb.GetTemplateRequest,
+) (*ttnpb.EndDeviceTemplate, error) {
+	if err := rights.RequireAuthentication(ctx); err != nil {
 		return nil, err
 	}
 	return dr.store.GetTemplate(req, nil)
 }
 
-func getDecoder(ctx context.Context, req *ttnpb.GetPayloadFormatterRequest, f func(store.GetCodecRequest) (*ttnpb.MessagePayloadDecoder, error)) (*ttnpb.MessagePayloadDecoder, error) {
+func getDecoder(
+	ctx context.Context,
+	req *ttnpb.GetPayloadFormatterRequest,
+	f func(store.GetCodecRequest) (*ttnpb.MessagePayloadDecoder, error),
+) (*ttnpb.MessagePayloadDecoder, error) {
 	if clusterauth.Authorized(ctx) != nil {
-		if err := requireApplication(ctx, req.GetApplicationIds()); err != nil {
+		if err := rights.RequireAuthentication(ctx); err != nil {
 			return nil, err
 		}
 	}
@@ -182,19 +197,28 @@ func getDecoder(ctx context.Context, req *ttnpb.GetPayloadFormatterRequest, f fu
 }
 
 // GetUplinkDecoder implements the ttnpb.DeviceRepositoryServer interface.
-func (dr *DeviceRepository) GetUplinkDecoder(ctx context.Context, req *ttnpb.GetPayloadFormatterRequest) (*ttnpb.MessagePayloadDecoder, error) {
+func (dr *DeviceRepository) GetUplinkDecoder(
+	ctx context.Context,
+	req *ttnpb.GetPayloadFormatterRequest,
+) (*ttnpb.MessagePayloadDecoder, error) {
 	return getDecoder(ctx, req, dr.store.GetUplinkDecoder)
 }
 
 // GetDownlinkDecoder implements the ttnpb.DeviceRepositoryServer interface.
-func (dr *DeviceRepository) GetDownlinkDecoder(ctx context.Context, req *ttnpb.GetPayloadFormatterRequest) (*ttnpb.MessagePayloadDecoder, error) {
+func (dr *DeviceRepository) GetDownlinkDecoder(
+	ctx context.Context,
+	req *ttnpb.GetPayloadFormatterRequest,
+) (*ttnpb.MessagePayloadDecoder, error) {
 	return getDecoder(ctx, req, dr.store.GetDownlinkDecoder)
 }
 
 // GetDownlinkEncoder implements the ttnpb.DeviceRepositoryServer interface.
-func (dr *DeviceRepository) GetDownlinkEncoder(ctx context.Context, req *ttnpb.GetPayloadFormatterRequest) (*ttnpb.MessagePayloadEncoder, error) {
+func (dr *DeviceRepository) GetDownlinkEncoder(
+	ctx context.Context,
+	req *ttnpb.GetPayloadFormatterRequest,
+) (*ttnpb.MessagePayloadEncoder, error) {
 	if clusterauth.Authorized(ctx) != nil {
-		if err := requireApplication(ctx, req.GetApplicationIds()); err != nil {
+		if err := rights.RequireAuthentication(ctx); err != nil {
 			return nil, err
 		}
 	}