+++ pkg/devicerepository/grpc_test.go
@@ -27,6 +27,7 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/devicerepository/store"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
 	mockis "go.thethings.network/lorawan-stack/v3/pkg/identityserver/mock"
+	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	"go.thethings.network/lorawan-stack/v3/pkg/rpcmetadata"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
@@ -89,12 +90,17 @@ func (s *mockStore) GetModels(req store.GetModelsRequest) (*store.GetModelsRespo
 }
 
 // GetEndDeviceProfiles implements store.
-func (s *mockStore) GetEndDeviceProfiles(req store.GetEndDeviceProfilesRequest) (*store.GetEndDeviceProfilesResponse, error) {
-	return nil, nil
+func (*mockStore) GetEndDeviceProfiles(
+	store.GetEndDeviceProfilesRequest,
+) (*store.GetEndDeviceProfilesResponse, error) {
+	return nil, nil //nolint:nilnil
 }
 
 // GetTemplate retrieves an end device template for an end device definition.
-func (s *mockStore) GetTemplate(req *ttnpb.GetTemplateRequest, _ *store.EndDeviceProfile) (*ttnpb.EndDeviceTemplate, error) {
+func (s *mockStore) GetTemplate(
+	req *ttnpb.GetTemplateRequest,
+	_ *store.EndDeviceProfile,
+) (*ttnpb.EndDeviceTemplate, error) {
 	s.lastVersionIDs = req.GetVersionIds()
 	profileIDs := req.GetEndDeviceProfileIds()
 	if profileIDs != nil {
@@ -127,40 +133,15 @@ func (s *mockStore) GetDownlinkEncoder(req store.GetCodecRequest) (*ttnpb.Messag
 }
 
 // Close closes the store.
-func (s *mockStore) Close() error {
+func (*mockStore) Close() error {
 	return nil
 }
 
 func TestGRPC(t *testing.T) {
-	ids := &ttnpb.EndDeviceVersionIdentifiers{
-		BrandId:         "brand",
-		ModelId:         "model",
-		FirmwareVersion: "1.0",
-		HardwareVersion: "1.0",
-		BandId:          "band",
-	}
+	ctx := log.NewContext(test.Context(), test.GetLogger(t))
 
-	registeredApplicationID := &ttnpb.ApplicationIdentifiers{
-		ApplicationId: "app1",
-	}
-	registeredApplicationKey := "key"
-	unregisteredApplicationKey := "unregisteredKey"
-	unregisteredApplicationID := ttnpb.ApplicationIdentifiers{
-		ApplicationId: "app2",
-	}
-	creds := grpc.PerRPCCredentials(rpcmetadata.MD{
-		AuthType:      "Bearer",
-		AuthValue:     registeredApplicationKey,
-		AllowInsecure: true,
-	})
-	unregisteredCreds := grpc.PerRPCCredentials(rpcmetadata.MD{
-		AuthType:      "Bearer",
-		AuthValue:     unregisteredApplicationKey,
-		AllowInsecure: true,
-	})
-	mockIS, mockISAddr, closeIS := mockis.New(test.Context())
+	mockIS, mockISAddr, closeIS := mockis.New(ctx)
 	defer closeIS()
-	mockIS.ApplicationRegistry().Add(test.Context(), registeredApplicationID, registeredApplicationKey, ttnpb.Right_RIGHT_APPLICATION_DEVICES_READ)
 
 	componentConfig := &component.Config{
 		ServiceBase: config.ServiceBase{
@@ -188,60 +169,60 @@ func TestGRPC(t *testing.T) {
 	componenttest.StartComponent(t, c)
 	defer c.Close()
 
-	mustHavePeer(test.Context(), c, ttnpb.ClusterRole_ENTITY_REGISTRY)
+	mustHavePeer(ctx, c, ttnpb.ClusterRole_ENTITY_REGISTRY)
 	cc := dr.LoopbackConn()
 	cl := ttnpb.NewDeviceRepositoryClient(cc)
+
+	creds := grpc.PerRPCCredentials(rpcmetadata.MD{
+		AuthType:      "Bearer",
+		AuthValue:     "MockValue",
+		AllowInsecure: true,
+	})
 
 	t.Run("Auth", func(t *testing.T) {
 		for _, atc := range []struct {
 			name    string
-			execute func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error)
+			execute func(opts ...grpc.CallOption) (interface{}, error)
 		}{
 			{
 				name: "ListBrands",
-				execute: func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error) {
-					response, err := cl.ListBrands(test.Context(), &ttnpb.ListEndDeviceBrandsRequest{
-						ApplicationIds: ids,
-					}, opts...)
+				execute: func(opts ...grpc.CallOption) (interface{}, error) {
+					response, err := cl.ListBrands(ctx, &ttnpb.ListEndDeviceBrandsRequest{}, opts...)
 					return response, err
 				},
 			},
 			{
 				name: "GetBrand",
-				execute: func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error) {
-					response, err := cl.GetBrand(test.Context(), &ttnpb.GetEndDeviceBrandRequest{
-						ApplicationIds: ids,
-						BrandId:        "brand1",
+				execute: func(opts ...grpc.CallOption) (interface{}, error) {
+					response, err := cl.GetBrand(ctx, &ttnpb.GetEndDeviceBrandRequest{
+						BrandId: "brand1",
 					}, opts...)
 					return response, err
 				},
 			},
 			{
 				name: "ListModels",
-				execute: func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error) {
-					response, err := cl.ListModels(test.Context(), &ttnpb.ListEndDeviceModelsRequest{
-						ApplicationIds: ids,
-						BrandId:        "brand1",
+				execute: func(opts ...grpc.CallOption) (interface{}, error) {
+					response, err := cl.ListModels(ctx, &ttnpb.ListEndDeviceModelsRequest{
+						BrandId: "brand1",
 					}, opts...)
 					return response, err
 				},
 			},
 			{
 				name: "GetModel",
-				execute: func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error) {
-					response, err := cl.GetModel(test.Context(), &ttnpb.GetEndDeviceModelRequest{
-						ApplicationIds: ids,
-						BrandId:        "brand1",
-						ModelId:        "model1",
+				execute: func(opts ...grpc.CallOption) (interface{}, error) {
+					response, err := cl.GetModel(ctx, &ttnpb.GetEndDeviceModelRequest{
+						BrandId: "brand1",
+						ModelId: "model1",
 					}, opts...)
 					return response, err
 				},
 			},
 			{
 				name: "GetTemplate",
-				execute: func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error) {
-					response, err := cl.GetTemplate(test.Context(), &ttnpb.GetTemplateRequest{
-						ApplicationIds: ids,
+				execute: func(opts ...grpc.CallOption) (interface{}, error) {
+					response, err := cl.GetTemplate(ctx, &ttnpb.GetTemplateRequest{
 						VersionIds: &ttnpb.EndDeviceVersionIdentifiers{
 							BrandId: "brand1",
 							ModelId: "model1",
@@ -252,9 +233,8 @@ func TestGRPC(t *testing.T) {
 			},
 			{
 				name: "GetUplinkDecoder",
-				execute: func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error) {
-					response, err := cl.GetUplinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-						ApplicationIds: ids,
+				execute: func(opts ...grpc.CallOption) (interface{}, error) {
+					response, err := cl.GetUplinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
 						VersionIds: &ttnpb.EndDeviceVersionIdentifiers{
 							BrandId: "brand1",
 							ModelId: "model1",
@@ -265,9 +245,8 @@ func TestGRPC(t *testing.T) {
 			},
 			{
 				name: "GetDownlinkDecoder",
-				execute: func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error) {
-					response, err := cl.GetDownlinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-						ApplicationIds: ids,
+				execute: func(opts ...grpc.CallOption) (interface{}, error) {
+					response, err := cl.GetDownlinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
 						VersionIds: &ttnpb.EndDeviceVersionIdentifiers{
 							BrandId: "brand1",
 							ModelId: "model1",
@@ -278,9 +257,8 @@ func TestGRPC(t *testing.T) {
 			},
 			{
 				name: "GetDownlinkEncoder",
-				execute: func(ids *ttnpb.ApplicationIdentifiers, opts ...grpc.CallOption) (interface{}, error) {
-					response, err := cl.GetDownlinkEncoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-						ApplicationIds: ids,
+				execute: func(opts ...grpc.CallOption) (interface{}, error) {
+					response, err := cl.GetDownlinkEncoder(ctx, &ttnpb.GetPayloadFormatterRequest{
 						VersionIds: &ttnpb.EndDeviceVersionIdentifiers{
 							BrandId: "brand1",
 							ModelId: "model1",
@@ -292,42 +270,36 @@ func TestGRPC(t *testing.T) {
 		} {
 			t.Run(atc.name, func(t *testing.T) {
 				for _, tc := range []struct {
-					name      string
-					ids       *ttnpb.ApplicationIdentifiers
-					opts      []grpc.CallOption
-					assertion func(err error) bool
+					name               string
+					opts               []grpc.CallOption
+					injectEntityAccess func()
+					assertion          func(err error) bool
 				}{
 					{
-						name:      "NoApplication",
-						assertion: errors.IsInvalidArgument,
-					},
-					{
-						name:      "NoKey",
-						ids:       registeredApplicationID,
+						name:      "no_creds",
 						assertion: errors.IsUnauthenticated,
 					},
 					{
-						name:      "UnregisteredKey",
-						ids:       registeredApplicationID,
-						opts:      []grpc.CallOption{unregisteredCreds},
-						assertion: errors.IsPermissionDenied,
-					},
-					{
-						name:      "UnregisteredApplication",
-						ids:       &unregisteredApplicationID,
-						assertion: errors.IsUnauthenticated,
-					},
-					{
-						name:      "UnregisteredApplicationWithKey",
-						ids:       &unregisteredApplicationID,
-						opts:      []grpc.CallOption{creds},
-						assertion: errors.IsPermissionDenied,
+						name:               "no_rights",
+						opts:               []grpc.CallOption{creds},
+						injectEntityAccess: func() { mockIS.EntityAccess().SetAuthInfoResponse(&ttnpb.AuthInfoResponse{}) },
+						assertion:          errors.IsUnauthenticated,
 					},
 				} {
 					t.Run(tc.name, func(t *testing.T) {
 						a := assertions.New(t)
-						response, err := atc.execute(tc.ids, tc.opts...)
+						defer func() {
+							mockIS.EntityAccess().SetAuthInfoResponse(nil)
+							mockIS.EntityAccess().SetErr(nil)
+						}()
+
+						if tc.injectEntityAccess != nil {
+							tc.injectEntityAccess()
+						}
+
+						response, err := atc.execute(tc.opts...)
 						a.So(response, should.BeNil)
+						t.Log(err)
 						a.So(tc.assertion(err), should.BeTrue)
 					})
 				}
@@ -335,17 +307,37 @@ func TestGRPC(t *testing.T) {
 		}
 	})
 
+	ids := &ttnpb.EndDeviceVersionIdentifiers{
+		BrandId:         "brand",
+		ModelId:         "model",
+		FirmwareVersion: "1.0",
+		HardwareVersion: "1.0",
+		BandId:          "band",
+	}
+
+	// Add rights to entity access mock.
+	mockIS.EntityAccess().SetAuthInfoResponse(&ttnpb.AuthInfoResponse{
+		AccessMethod: &ttnpb.AuthInfoResponse_ApiKey{
+			ApiKey: &ttnpb.AuthInfoResponse_APIKeyAccess{
+				ApiKey: &ttnpb.APIKey{
+					// The specific right here doesn't have an affect since the
+					// validation that happens only checks if its nil.
+					Rights: []ttnpb.Right{ttnpb.Right_RIGHT_USER_ALL},
+				},
+			},
+		},
+	})
+
 	t.Run("ListBrands", func(t *testing.T) {
 		t.Run("Request", func(t *testing.T) {
 			a := assertions.New(t)
 
-			_, err := cl.ListBrands(test.Context(), &ttnpb.ListEndDeviceBrandsRequest{
-				ApplicationIds: registeredApplicationID,
-				Limit:          100,
-				Page:           2,
-				OrderBy:        "brand_id",
-				Search:         "query string",
-				FieldMask:      ttnpb.FieldMask("lora_alliance_vendor_id"),
+			_, err := cl.ListBrands(ctx, &ttnpb.ListEndDeviceBrandsRequest{
+				Limit:     100,
+				Page:      2,
+				OrderBy:   "brand_id",
+				Search:    "query string",
+				FieldMask: ttnpb.FieldMask("lora_alliance_vendor_id"),
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(st.lastGetBrandsRequest, should.Resemble, store.GetBrandsRequest{
@@ -361,9 +353,7 @@ func TestGRPC(t *testing.T) {
 		t.Run("StoreError", func(t *testing.T) {
 			a := assertions.New(t)
 			st.err = fmt.Errorf("store error")
-			brands, err := cl.ListBrands(test.Context(), &ttnpb.ListEndDeviceBrandsRequest{
-				ApplicationIds: registeredApplicationID,
-			}, creds)
+			brands, err := cl.ListBrands(ctx, &ttnpb.ListEndDeviceBrandsRequest{}, creds)
 			a.So(brands, should.BeNil)
 			a.So(err.Error(), should.ContainSubstring, st.err.Error())
 		})
@@ -376,16 +366,12 @@ func TestGRPC(t *testing.T) {
 					PrivateEnterpriseNumber: 100,
 					Logo:                    "item.png",
 				},
-				{
-					BrandId: "brand2",
-				},
+				{BrandId: "brand2"},
 			}
 			st.err = nil
 
 			responseHeaders := metadata.MD{}
-			brands, err := cl.ListBrands(test.Context(), &ttnpb.ListEndDeviceBrandsRequest{
-				ApplicationIds: registeredApplicationID,
-			}, grpc.Header(&responseHeaders), creds)
+			brands, err := cl.ListBrands(ctx, &ttnpb.ListEndDeviceBrandsRequest{}, grpc.Header(&responseHeaders), creds)
 			a.So(err, should.BeNil)
 			a.So(brands, should.Resemble, &ttnpb.ListEndDeviceBrandsResponse{
 				Brands: []*ttnpb.EndDeviceBrand{
@@ -409,10 +395,9 @@ func TestGRPC(t *testing.T) {
 		t.Run("Request", func(t *testing.T) {
 			a := assertions.New(t)
 
-			_, err := cl.GetBrand(test.Context(), &ttnpb.GetEndDeviceBrandRequest{
-				ApplicationIds: registeredApplicationID,
-				BrandId:        "brand1",
-				FieldMask:      ttnpb.FieldMask("lora_alliance_vendor_id"),
+			_, err := cl.GetBrand(ctx, &ttnpb.GetEndDeviceBrandRequest{
+				BrandId:   "brand1",
+				FieldMask: ttnpb.FieldMask("lora_alliance_vendor_id"),
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(st.lastGetBrandsRequest, should.Resemble, store.GetBrandsRequest{
@@ -425,9 +410,7 @@ func TestGRPC(t *testing.T) {
 		t.Run("StoreError", func(t *testing.T) {
 			a := assertions.New(t)
 			st.err = fmt.Errorf("store error")
-			brands, err := cl.ListBrands(test.Context(), &ttnpb.ListEndDeviceBrandsRequest{
-				ApplicationIds: registeredApplicationID,
-			}, creds)
+			brands, err := cl.ListBrands(ctx, &ttnpb.ListEndDeviceBrandsRequest{}, creds)
 			a.So(brands, should.BeNil)
 			a.So(err.Error(), should.ContainSubstring, st.err.Error())
 		})
@@ -443,9 +426,8 @@ func TestGRPC(t *testing.T) {
 			}
 			st.err = nil
 
-			brand, err := cl.GetBrand(test.Context(), &ttnpb.GetEndDeviceBrandRequest{
-				ApplicationIds: registeredApplicationID,
-				BrandId:        "brand1",
+			brand, err := cl.GetBrand(ctx, &ttnpb.GetEndDeviceBrandRequest{
+				BrandId: "brand1",
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(brand, should.Resemble, &ttnpb.EndDeviceBrand{
@@ -460,14 +442,13 @@ func TestGRPC(t *testing.T) {
 		t.Run("Request", func(t *testing.T) {
 			a := assertions.New(t)
 
-			_, err := cl.ListModels(test.Context(), &ttnpb.ListEndDeviceModelsRequest{
-				ApplicationIds: registeredApplicationID,
-				BrandId:        "brand1",
-				Limit:          100,
-				Page:           2,
-				OrderBy:        "brand_id",
-				Search:         "query string",
-				FieldMask:      ttnpb.FieldMask("firmware_versions"),
+			_, err := cl.ListModels(ctx, &ttnpb.ListEndDeviceModelsRequest{
+				BrandId:   "brand1",
+				Limit:     100,
+				Page:      2,
+				OrderBy:   "brand_id",
+				Search:    "query string",
+				FieldMask: ttnpb.FieldMask("firmware_versions"),
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(st.lastGetModelsRequest, should.Resemble, store.GetModelsRequest{
@@ -484,9 +465,7 @@ func TestGRPC(t *testing.T) {
 		t.Run("StoreError", func(t *testing.T) {
 			a := assertions.New(t)
 			st.err = fmt.Errorf("store error")
-			res, err := cl.ListModels(test.Context(), &ttnpb.ListEndDeviceModelsRequest{
-				ApplicationIds: registeredApplicationID,
-			}, creds)
+			res, err := cl.ListModels(ctx, &ttnpb.ListEndDeviceModelsRequest{}, creds)
 			a.So(res, should.BeNil)
 			a.So(err.Error(), should.ContainSubstring, st.err.Error())
 		})
@@ -510,9 +489,7 @@ func TestGRPC(t *testing.T) {
 			st.err = nil
 
 			responseHeaders := metadata.MD{}
-			brands, err := cl.ListModels(test.Context(), &ttnpb.ListEndDeviceModelsRequest{
-				ApplicationIds: registeredApplicationID,
-			}, grpc.Header(&responseHeaders), creds)
+			brands, err := cl.ListModels(ctx, &ttnpb.ListEndDeviceModelsRequest{}, grpc.Header(&responseHeaders), creds)
 			a.So(err, should.BeNil)
 			a.So(brands, should.Resemble, &ttnpb.ListEndDeviceModelsResponse{
 				Models: []*ttnpb.EndDeviceModel{
@@ -540,11 +517,10 @@ func TestGRPC(t *testing.T) {
 		t.Run("Request", func(t *testing.T) {
 			a := assertions.New(t)
 
-			_, err := cl.GetModel(test.Context(), &ttnpb.GetEndDeviceModelRequest{
-				ApplicationIds: registeredApplicationID,
-				BrandId:        "brand1",
-				ModelId:        "model1",
-				FieldMask:      ttnpb.FieldMask("firmware_versions"),
+			_, err := cl.GetModel(ctx, &ttnpb.GetEndDeviceModelRequest{
+				BrandId:   "brand1",
+				ModelId:   "model1",
+				FieldMask: ttnpb.FieldMask("firmware_versions"),
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(st.lastGetModelsRequest, should.Resemble, store.GetModelsRequest{
@@ -558,9 +534,7 @@ func TestGRPC(t *testing.T) {
 		t.Run("StoreError", func(t *testing.T) {
 			a := assertions.New(t)
 			st.err = fmt.Errorf("store error")
-			models, err := cl.ListModels(test.Context(), &ttnpb.ListEndDeviceModelsRequest{
-				ApplicationIds: registeredApplicationID,
-			}, creds)
+			models, err := cl.ListModels(ctx, &ttnpb.ListEndDeviceModelsRequest{}, creds)
 			a.So(models, should.BeNil)
 			a.So(err.Error(), should.ContainSubstring, st.err.Error())
 		})
@@ -579,10 +553,9 @@ func TestGRPC(t *testing.T) {
 			}
 			st.err = nil
 
-			model, err := cl.GetModel(test.Context(), &ttnpb.GetEndDeviceModelRequest{
-				ApplicationIds: registeredApplicationID,
-				BrandId:        "brand1",
-				ModelId:        "model1",
+			model, err := cl.GetModel(ctx, &ttnpb.GetEndDeviceModelRequest{
+				BrandId: "brand1",
+				ModelId: "model1",
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(model, should.Resemble, &ttnpb.EndDeviceModel{
@@ -606,9 +579,8 @@ func TestGRPC(t *testing.T) {
 
 		t.Run("Request", func(t *testing.T) {
 			a := assertions.New(t)
-			_, err := cl.GetTemplate(test.Context(), &ttnpb.GetTemplateRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			_, err := cl.GetTemplate(ctx, &ttnpb.GetTemplateRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(st.lastVersionIDs, should.Resemble, ids)
@@ -617,9 +589,8 @@ func TestGRPC(t *testing.T) {
 		t.Run("StoreError", func(t *testing.T) {
 			a := assertions.New(t)
 			st.err = fmt.Errorf("store error")
-			models, err := cl.GetTemplate(test.Context(), &ttnpb.GetTemplateRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			models, err := cl.GetTemplate(ctx, &ttnpb.GetTemplateRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(models, should.BeNil)
 			a.So(err.Error(), should.ContainSubstring, st.err.Error())
@@ -629,9 +600,8 @@ func TestGRPC(t *testing.T) {
 			a := assertions.New(t)
 			st.err = nil
 
-			template, err := cl.GetTemplate(test.Context(), &ttnpb.GetTemplateRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			template, err := cl.GetTemplate(ctx, &ttnpb.GetTemplateRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(template, should.Resemble, st.template)
@@ -654,8 +624,7 @@ func TestGRPC(t *testing.T) {
 
 		t.Run("Request", func(t *testing.T) {
 			a := assertions.New(t)
-			_, err := cl.GetTemplate(test.Context(), &ttnpb.GetTemplateRequest{
-				ApplicationIds:      registeredApplicationID,
+			_, err := cl.GetTemplate(ctx, &ttnpb.GetTemplateRequest{
 				EndDeviceProfileIds: profileIDs,
 			}, creds)
 			a.So(err, should.BeNil)
@@ -667,8 +636,7 @@ func TestGRPC(t *testing.T) {
 		t.Run("StoreError", func(t *testing.T) {
 			a := assertions.New(t)
 			st.err = fmt.Errorf("store error")
-			models, err := cl.GetTemplate(test.Context(), &ttnpb.GetTemplateRequest{
-				ApplicationIds:      registeredApplicationID,
+			models, err := cl.GetTemplate(ctx, &ttnpb.GetTemplateRequest{
 				EndDeviceProfileIds: profileIDs,
 			}, creds)
 			a.So(models, should.BeNil)
@@ -679,8 +647,7 @@ func TestGRPC(t *testing.T) {
 			a := assertions.New(t)
 			st.err = nil
 
-			template, err := cl.GetTemplate(test.Context(), &ttnpb.GetTemplateRequest{
-				ApplicationIds:      registeredApplicationID,
+			template, err := cl.GetTemplate(ctx, &ttnpb.GetTemplateRequest{
 				EndDeviceProfileIds: profileIDs,
 			}, creds)
 			a.So(err, should.BeNil)
@@ -694,9 +661,8 @@ func TestGRPC(t *testing.T) {
 			st.lastVersionIDs = nil
 			st.lastCodecPaths = nil
 			st.err = fmt.Errorf("store error")
-			c, err := cl.GetUplinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			c, err := cl.GetUplinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(c, should.BeNil)
 			a.So(err.Error(), should.ContainSubstring, st.err.Error())
@@ -711,26 +677,22 @@ func TestGRPC(t *testing.T) {
 			}
 			st.err = nil
 
-			c, err := cl.GetUplinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			c, err := cl.GetUplinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(c, should.Resemble, st.uplinkDecoder)
 		})
-
 		t.Run("ClusterAuth", func(t *testing.T) {
-			codec, err := cl.GetUplinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			codec, err := cl.GetUplinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			})
 			a := assertions.New(t)
 			a.So(errors.IsUnauthenticated(err), should.BeTrue)
 			a.So(codec, should.BeNil)
 
-			_, err = cl.GetUplinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			_, err = cl.GetUplinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, c.WithClusterAuth())
 			a.So(err, should.BeNil)
 		})
@@ -742,9 +704,8 @@ func TestGRPC(t *testing.T) {
 			st.lastVersionIDs = nil
 			st.lastCodecPaths = nil
 			st.err = fmt.Errorf("store error")
-			c, err := cl.GetDownlinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			c, err := cl.GetDownlinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(c, should.BeNil)
 			a.So(err.Error(), should.ContainSubstring, st.err.Error())
@@ -759,26 +720,22 @@ func TestGRPC(t *testing.T) {
 			}
 			st.err = nil
 
-			c, err := cl.GetDownlinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			c, err := cl.GetDownlinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(c, should.Resemble, st.downlinkDecoder)
 		})
-
 		t.Run("ClusterAuth", func(t *testing.T) {
-			codec, err := cl.GetDownlinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			codec, err := cl.GetDownlinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			})
 			a := assertions.New(t)
 			a.So(errors.IsUnauthenticated(err), should.BeTrue)
 			a.So(codec, should.BeNil)
 
-			_, err = cl.GetDownlinkDecoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			_, err = cl.GetDownlinkDecoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, c.WithClusterAuth())
 			a.So(err, should.BeNil)
 		})
@@ -790,9 +747,8 @@ func TestGRPC(t *testing.T) {
 			st.lastVersionIDs = nil
 			st.lastCodecPaths = nil
 			st.err = fmt.Errorf("store error")
-			c, err := cl.GetDownlinkEncoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			c, err := cl.GetDownlinkEncoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(c, should.BeNil)
 			a.So(err.Error(), should.ContainSubstring, st.err.Error())
@@ -807,26 +763,22 @@ func TestGRPC(t *testing.T) {
 			}
 			st.err = nil
 
-			c, err := cl.GetDownlinkEncoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			c, err := cl.GetDownlinkEncoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, creds)
 			a.So(err, should.BeNil)
 			a.So(c, should.Resemble, st.downlinkEncoder)
 		})
-
 		t.Run("ClusterAuth", func(t *testing.T) {
-			codec, err := cl.GetDownlinkEncoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			codec, err := cl.GetDownlinkEncoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			})
 			a := assertions.New(t)
 			a.So(errors.IsUnauthenticated(err), should.BeTrue)
 			a.So(codec, should.BeNil)
 
-			_, err = cl.GetDownlinkEncoder(test.Context(), &ttnpb.GetPayloadFormatterRequest{
-				ApplicationIds: registeredApplicationID,
-				VersionIds:     ids,
+			_, err = cl.GetDownlinkEncoder(ctx, &ttnpb.GetPayloadFormatterRequest{
+				VersionIds: ids,
 			}, c.WithClusterAuth())
 			a.So(err, should.BeNil)
 		})